renamed standard to singlecluster
On-behalf-of: @SAP a.shcherbatiuk@sap.com
